Move mock-xhr helper to test/helpers to avoid Jest picking it up
Jest's testMatch pattern `**/test/*.[jt]s?(x)` matches any .ts file
directly inside a test/ directory. Moving mock-xhr.ts to test/helpers/
avoids it being treated as a test suite (which fails because it
contains no tests) while still keeping it excluded from the TypeScript
production build.
